Several ports fixes

+++ mbcico/hydra.c
@@ -902,7 +902,7 @@ int hydra_batch(int role, file_list *to_send)
 		} else {
 		    if (to_send) {
 			txlen = sprintf(txbuf, "%08lx%08lx%08lx%08lx%08lx",
-					mtime2sl(txstat.st_mtime+(txstat.st_mtime%2)),
+					(long)mtime2sl(txstat.st_mtime+(txstat.st_mtime%2)),
 					(long)(txstat.st_size), 0UL, 0UL, 0UL);
 
 			/*
